Move backfill; Fix base to use zb dataclass;  Fix iterable wrapper to not rely on abstractbackedlist
