+++ src/zepben/ewb/boilerplate/collections/abstract_backed_collections.py
@@ -3,7 +3,7 @@
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 from abc import ABC, abstractmethod
-from typing import Collection, Protocol, Iterable, Generic, Iterator, Callable, Any, Sequence, overload, TypeVar
+from typing import Collection, Iterable, Generic, Iterator, Callable, Any, Sequence, overload, TypeVar
 
 
 T = TypeVar("T")
