+++ src/zepben/ewb/model/cim/iec61970/base/wires/energy_source.py
@@ -14,7 +14,8 @@ from typing_extensions import deprecated
 
 from zepben.ewb import Alias
 from zepben.ewb.boilerplate.dataclass_base import zb_dataclass
-from zepben.ewb.boilerplate.collections.mrid_list import MridCollection, LazyMridList, Backfill
+from zepben.ewb.boilerplate.collections.mrid_list import MridCollection, LazyMridList
+from zepben.ewb.boilerplate.backfill import Backfill
 from zepben.ewb.model.cim.iec61970.base.wires.energy_connection import EnergyConnection
 from zepben.ewb.model.cim.iec61970.base.wires.energy_source_phase import EnergySourcePhase
