+++ src/zepben/ewb/model/cim/iec61970/base/wires/energy_source_phase.py
@@ -10,7 +10,7 @@ from typing import Optional, TYPE_CHECKING
 
 from typing_extensions import deprecated
 
-from zepben.ewb.boilerplate.collections.mrid_list import internal
+from zepben.ewb.boilerplate.backfill import internal
 from zepben.ewb.model.cim.iec61970.base.core.power_system_resource import PowerSystemResource
 from zepben.ewb.model.cim.iec61970.base.wires.single_phase_kind import SinglePhaseKind
 from zepben.ewb.boilerplate.dataclass_base import zb_dataclass
